src/requirements.txt
# SPC Framework Requirements

# Core dependencies
numpy>=1.21.0
PyYAML>=5.4.0
networkx>=2.6.0
scikit-learn>=1.0.0

# Optional ML dependencies (for enhanced functionality)
tensorflow>=2.8.0; python_version>='3.7'
torch>=1.9.0; python_version>='3.7'

# Utility
python-dateutil>=2.8.0

# Development (optional)
pytest>=6.2.0
pytest-cov>=2.12.0
